Add to the sum when new card is clicked
Create a New Card button and its function that logs out the message 'Drawing a new card from the deck' when the button is clicked. Also, create a variabe named card and add to the sum variable using += operator and call the starGame() function. By the time the code runs and reaches the startGame() function the sum variable would be updated to include the New Card. This new card would be rendered along side the first tw0 cards.