Create a for loop function to display all cards instead of just two
Create a for loop function that displays all the cards available rather than just the two. We use the length property as follows: cards.length, rather than a fixed number, to enable us to update the array list dynamically and still render the entire list. We render the list by attaching [i] to the array name variable, cards as follows: cards[i].